Exceptions

+++ controllers/artistController.php
@@ -14,25 +14,45 @@ class ArtistController
 
     public function index()
     {
-        include(ROOT. "views/createEventForm.php");
+        //include(ROOT. "views/createEventForm.php");
+        include(ROOT. "views/artistForm.php");
     }
   
     public function store($nombre)
     {
-        $artist = new Artist($nombre);
-        $this->dao->insert($artist);
+        $flag = $this->searchInDatabase($nombre);
+        if(!$flag){
+            $artist = new Artist($nombre);
+            $ultimoID=$this->dao->insert($artist);
+            return $ultimoID;
+        }
+        else{
+            throw new \Exception ('El artista ya existe');
+        }
     }
 
     public function delete($nombre)
     {
-        $artist = new Artist($nombre);
-        $this->dao->delete($artist);
+        $flag = $this->searchInDatabase($nombre);
+        if($flag){
+            $artist = new Artist($nombre);
+            $this->dao->delete($artist);
+        }
+        else{
+            throw new \Exception ('Ha ocurrido un error'); 
+        }
     }
 
     public function update($nombre, $nuevoDato)
     {
-        $artist = new Artist($nombre);
-        $this->dao->update($artist, $nuevoDato);
+        $flag = $this->searchInDatabase($nombre);
+        if($flag){
+            $artist = new Artist($nombre);
+            $this->dao->update($artist, $nuevoDato);
+        }
+        else{
+            throw new \Exception ('Ha ocurrido un error');
+        }
     }
 
     public function retride(){
@@ -44,4 +64,17 @@ class ArtistController
         $id = $this->dao->getIdByName($nombre);
     }
     
+    public function searchInDatabase($nombre){
+        $list = $this->retride();
+        $flag = false;
+        if($list){
+            foreach ($list as $key => $value) {
+                if($value === $nombre){
+                    $flag = true;
+                }
+            }
+        }
+        return $flag;
+    }
+
 }
