+++ controllers/eventController.php
@@ -40,10 +40,6 @@ class EventController{
 
         $parametersEvent = $this->getEventData();
         $this->getCalendarData($parametersEvent);
-
-        /*if($_POST){
-            $parametersCalendar = $_POST;
-        }*/
     }
 
     public function index2(){
@@ -59,36 +55,71 @@ class EventController{
 
    public function store()
     {
-        if($_POST){
-            $counter = 0;
-            $event = new Event($_SESSION['eventData']['name'], $_SESSION['eventData']['category']);
-            foreach ($_POST as $key => $value) {
-                $eventDate = $_SESSION['eventData']['eventDates'][$counter];
-                $eventPlace = $_SESSION['eventData']['eventPlace'];
-                $event->setCalendar($eventDate, $eventPlace, $value);
-                $counter++;
+        $flag = $this->searchInDatabase($nombre);
+        if(!$flag){
+            if($_POST){
+                $counter = 0;
+                $event = new Event($_SESSION['eventData']['name'], $_SESSION['eventData']['category']);
+                foreach ($_POST as $key => $value) {
+                    $eventDate = $_SESSION['eventData']['eventDates'][$counter];
+                    $eventPlace = $_SESSION['eventData']['eventPlace'];
+                    $event->setCalendar($eventDate, $eventPlace, $value);
+                    $counter++;
+                }
+                $this->dao->insert($event);
+                $calendarControl = new CalendarController();
+                $calendarControl->store($event);
             }
-            $this->dao->insert($event);
-            $calendarControl = new CalendarController();
-            $calendarControl->store($event);
+        }
+        else{
+            throw new \Exception ('El evento ya existe');
         }
     }
 
     public function delete($nombre)
     {
-        $event = new Event($nombre);
-        $this->dao->delete($event);
+        $flag = $this->searchInDatabase($nombre);
+        if($flag){
+            $event = new Event($nombre);
+            $this->dao->delete($event);
+        }
+        else{
+            throw new \Exception ('Ha ocurrido un error'); 
+        }
     }
 
     public function update($nombre, $nuevoDato)
     {
-        $event = new Event($nombre);
-        $this->dao->update($event, $nuevoDato);
+        $flag = $this->searchInDatabase($nombre);
+        if($flag){
+            $event = new Event($nombre);
+            $this->dao->update($event, $nuevoDato);
+        }
+        else{
+            throw new \Exception ('Ha ocurrido un error'); 
+        }
     }
 
     public function retride(){
         $list=$this->dao->retride();
         return $list;
+    }
+
+    public function getIdByName($nombre){
+        $id = $this->dao->getIdByName($nombre);
+    }
+
+    public function searchInDatabase($nombre){
+        $list = $this->retride();
+        $flag = false;
+        if($list){
+            foreach ($list as $key => $value) {
+                if($value === $nombre){
+                    $flag = true;
+                }
+            }
+        }
+        return $flag;
     }
 
 }
